agents/scanner/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
azure-ai-projects>=2.0.0b3
azure-identity>=1.25.0
